Let ContentStreamRequestHandler respect Exif orientation
