+++ picasso/src/main/java/com/squareup/picasso/ContentStreamRequestHandler.java
@@ -36,7 +36,7 @@ class ContentStreamRequestHandler extends RequestHandler {
   }
 
   @Override public Result load(Request request, int networkPolicy) throws IOException {
-    return new Result(getInputStream(request), DISK);
+    return new Result(null, getInputStream(request), DISK, Utils.getExifOrientation(context.getContentResolver(), request.uri));
   }
 
   InputStream getInputStream(Request request) throws FileNotFoundException {
